+++ app/build.gradle
@@ -43,4 +43,5 @@ android {
 
 dependencies {
     implementation 'com.android.support.constraint:constraint-layout:+'
+    implementation 'com.android.support:appcompat-v7:28.0.0'
 }
