+++ app/src/main/java/com/buildmyapp/android/BottomNavFragments/BottomNavigation.java
@@ -4,8 +4,6 @@ import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.design.widget.BottomNavigationView;
 import android.support.v4.app.Fragment;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.AppCompatActivity;
 import android.view.MenuItem;
 
@@ -27,11 +25,6 @@ public class BottomNavigation extends AppCompatActivity  {
         bottomNav.setOnNavigationItemSelectedListener(navListener);
 
 
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
-       // ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
-               // this, drawer,, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
-      //  drawer.addDrawerListener(toggle);
-       // toggle.syncState();
 
         getSupportFragmentManager().beginTransaction().replace(R.id.fragment_container,
                 new HomeFragment()).commit();
@@ -56,10 +49,7 @@ public class BottomNavigation extends AppCompatActivity  {
                         case R.id.navigationMyDev:
                             selectedFragment = new DevelopersFragment();
                             break;
-                        case  R.id.navigationMenu:
-                            DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
-                            drawer.openDrawer(GravityCompat.START);
-                            return true;
+
                     }
 
                     getSupportFragmentManager().beginTransaction().replace(R.id.fragment_container,
