+++ app/src/main/java/com/buildmyapp/android/BottomNavFragments/DevelopersFragment.java
@@ -6,15 +6,13 @@ import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 
 import com.buildmyapp.android.R;
-import com.buildmyapp.android.bio;
+import com.buildmyapp.android.innerActivity.bio;
 
 public class DevelopersFragment extends Fragment {
