--- app/src/main/java/com/buildmyapp/android/BottomNavigationBehavior.java
@@ -1,50 +1,0 @@
-package com.buildmyapp.android;
-
-
-
-import android.content.Context;
-import android.support.design.widget.BottomNavigationView;
-import android.support.design.widget.CoordinatorLayout;
-import android.support.v4.view.ViewCompat;
-import android.util.AttributeSet;
-import android.view.View;
-import android.widget.FrameLayout;
-
-public class BottomNavigationBehavior extends CoordinatorLayout.Behavior<BottomNavigationView> {
-
-    public BottomNavigationBehavior() {
-        super();
-    }
-
-    public BottomNavigationBehavior(Context context, AttributeSet attrs) {
-        super(context, attrs);
-    }
-
-    @Override
-    public boolean layoutDependsOn(CoordinatorLayout parent, BottomNavigationView child, View dependency) {
-        boolean dependsOn = dependency instanceof FrameLayout;
-        return dependsOn;
-    }
-
-    @Override
-    public boolean onStartNestedScroll(CoordinatorLayout coordinatorLayout, BottomNavigationView child, View directTargetChild, View target, int nestedScrollAxes) {
-        return nestedScrollAxes == ViewCompat.SCROLL_AXIS_VERTICAL;
-    }
-
-    @Override
-    public void onNestedPreScroll(CoordinatorLayout coordinatorLayout, BottomNavigationView child, View target, int dx, int dy, int[] consumed) {
-        if (dy < 0) {
-            showBottomNavigationView(child);
-        } else if (dy > 0) {
-            hideBottomNavigationView(child);
-        }
-    }
-
-    private void hideBottomNavigationView(BottomNavigationView view) {
-        view.animate().translationY(view.getHeight());
-    }
-
-    private void showBottomNavigationView(BottomNavigationView view) {
-        view.animate().translationY(0);
-    }
-}
