--- app/src/main/java/com/buildmyapp/android/Fragments/MenuFragment.java
@@ -1,20 +1,0 @@
-package com.buildmyapp.android.Fragments;
-
-import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import com.buildmyapp.android.R;
-
-public class MenuFragment extends Fragment {
-
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.drawer_behavior, container, false);
-    }
-}
