--- app/src/main/java/com/buildmyapp/android/JsonTestApi.java
@@ -1,14 +1,0 @@
-package com.buildmyapp.android;
-
-import com.buildmyapp.android.model.Users;
-
-import java.util.List;
-
-import retrofit2.Call;
-import retrofit2.http.GET;
-
-public interface JsonTestApi {
-
-    @GET("api/users")
-    Call<List<Users>> getJson();
-}
