--- app/src/main/java/com/buildmyapp/android/Plan.java
@@ -1,13 +1,0 @@
-package com.buildmyapp.android;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class Plan extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_plan);
-    }
-}
