+++ app/src/main/java/com/buildmyapp/android/TestConnection.java
@@ -4,6 +4,7 @@ import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.widget.TextView;
 
+import com.buildmyapp.android.api.JsonTestApi;
 import com.buildmyapp.android.model.Users;
 
 import java.util.List;
