+++ app/src/main/java/com/buildmyapp/android/api/JsonDevApi.java
@@ -1,0 +1,14 @@
+package com.buildmyapp.android.api;
+
+import com.buildmyapp.android.model.Developer;
+
+import java.util.List;
+
+import retrofit2.Call;
+import retrofit2.http.GET;
+
+public interface JsonDevApi {
+
+    @GET("api/Developer")
+    Call<List<Developer>> getJson();
+}
