+++ app/src/main/java/com/buildmyapp/android/innerActivity/MessageSeller.java
@@ -1,0 +1,15 @@
+package com.buildmyapp.android.innerActivity;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+import com.buildmyapp.android.R;
+
+public class MessageSeller extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_message_seller);
+    }
+}
