+++ app/src/main/res/layout/activity_about_dev.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".aboutDev">
+    tools:context=".mainScreenActivity.aboutDev">
 
     <ImageView
         android:id="@+id/imageView3"
