+++ app/src/main/res/layout/activity_bio.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".bio">
+    tools:context=".innerActivity.bio">
 
 
     <android.support.v7.widget.CardView
