+++ app/src/main/res/layout/activity_message_seller.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MessageSeller">
+    tools:context=".innerActivity.MessageSeller">
 
     <android.support.v7.widget.CardView
         android:id="@+id/cardView_master_message"
