--- app/src/main/res/layout/plan.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-
-    <TextView
-        android:id="@+id/text_plan"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Developer's Plan"
-        tools:layout_editor_absoluteX="135dp"
-        tools:layout_editor_absoluteY="36dp" />
-</android.support.constraint.ConstraintLayout>
