+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -2,19 +2,15 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item
-        android:id="@+id/navigationMenu"
-        android:icon="@android:drawable/ic_menu_sort_by_size"
-        android:title="Menu" />
+        android:id="@+id/navigationHome"
+        android:icon="@drawable/ic_home"
+        android:title="Home" />
 
     <item
         android:id="@+id/navigationMyDev"
         android:icon="@drawable/ic_effort"
         android:title="Developers" />
 
-    <item
-        android:id="@+id/navigationHome"
-        android:icon="@drawable/ic_home"
-        android:title="Home" />
 
     <item
         android:id="@+id/navigationSearch"
@@ -24,6 +20,8 @@
     <item
         android:id="@+id/navigationMyProfile"
         android:icon="@drawable/ic_user"
-        android:title="User" />
+        android:title="Profile" />
+
+
 
 </menu>
